backend/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
beautifulsoup4==4.14.2
black==26.1.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
distlib==0.4.0
Django==6.0.1
django-ninja==1.5.3
dnspython==2.8.0
email-validator==2.3.0
filelock==3.20.3
flake8==7.3.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
identify==2.6.16
idna==3.11
iniconfig==2.1.0
isort==7.0.0
mccabe==0.7.0
msgpack==1.0.8
mypy_extensions==1.1.0
nodeenv==1.10.0
numpy==2.3.2
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pytest==8.4.2
python-can==4.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
virtualenv==20.36.1
wrapt==1.16.0
